Remove the broken link to the gmailFilters.xml file (#85)
I made a small change to make it more clear that the gmailFitlers.xml is not expected to be inside the repository in GitHub. That's the file I care the most about and the first link I clicked. The fact that it was broken made me think the project was broken or abandoned. Then I was about to report the issue and GitHub showed me the issue already existed, it was closed and this was the intended behaviour. I think these chanes reflect that a little bit better.

Incidentally, I think it would be good to have a way to provide the xml without needing to install Ruby.